# internal/spec/openapi.yaml
openapi: 3.1.0
info:
  title: hello
  version: 1.1.0
  description: Wake-on-LAN registry with live presence. Send magic packets, track
    devices, and subscribe to status changes over WebSocket.
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /devices:
    get:
      tags: [Devices]
      summary: List devices
      description: Filtered, sorted and paginated. The envelope always carries the total.
      operationId: listDevices
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/DeviceStatus'
        - name: search
          in: query
          required: false
          description: Case-insensitive substring over id, name, MAC and IP
          schema:
            type: string
        - name: tag
          in: query
          required: false
          schema:
            type: string
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [name, status, last_seen, wake_count]
            default: name
        - name: order
          in: query
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: asc
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
      responses:
        '200':
          description: Current registry snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
    post:
      tags: [Devices]
      summary: Register a device
      operationId: createDevice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceInput'
            examples:
              withPing:
                summary: Monitored host
                value:
                  id: pc-salon
                  name: PC Salon
                  mac: 00:11:22:33:44:55
                  ip: 192.168.1.50
                  ping_enabled: true
      responses:
        '201':
          description: Device registered, echo of the stored representation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '409':
          $ref: '#/components/responses/Conflict'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
    delete:
      tags: [Devices]
      summary: Delete several devices
      description: Comma-separated ids. Unknown ids are reported, not fatal.
      operationId: bulkDeleteDevices
      parameters:
        - name: ids
          in: query
          required: true
          schema:
            type: string
            example: pc-salon,pc-bureau
      responses:
        '200':
          description: Per-id report
          content:
            application/json:
              schema:
                type: object
                required: [deleted, not_found]
                properties:
                  deleted:
                    type: array
                    items:
                      type: string
                  not_found:
                    type: array
                    items:
                      type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
  /devices/{id}:
    parameters:
      - $ref: '#/components/parameters/DeviceId'
    get:
      tags: [Devices]
      summary: Get a device
      operationId: getDevice
      responses:
        '200':
          description: Stored representation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Devices]
      summary: Replace a device
      description: Full replacement. The path id wins over any id in the body.
        Omitted fields reset to defaults (status defaults to unknown).
      operationId: updateDevice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceInput'
      responses:
        '200':
          description: Stored representation after replacement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
    patch:
      tags: [Devices]
      summary: Patch a device
      description: JSON merge patch, only the keys present are changed. The id is immutable here.
      operationId: patchDevice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              properties:
                name:
                  type: string
                mac:
                  type: string
                  pattern: '^([0-9A-Fa-f]{2}[:-]){5}([0-9A-Fa-f]{2})$'
                ip:
                  type: string
                  format: ipv4
                status:
                  $ref: '#/components/schemas/DeviceStatus'
                ping_enabled:
                  type: boolean
                notes:
                  type: string
                  maxLength: 500
                tags:
                  type: array
                  maxItems: 10
                  items:
                    type: string
              additionalProperties: false
      responses:
        '200':
          description: Stored representation after patch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
    delete:
      tags: [Devices]
      summary: Delete a device
      operationId: deleteDevice
      responses:
        '204':
          description: Deleted, empty body
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
...
    DeviceId:
      name: id
      in: path
      required: true
      description: Registry identifier, trimmed on write
      schema:
        type: string
        minLength: 1
        example: pc-salon
  schemas:
    Device:
      type: object
      required: [id, name, mac, status, ping_enabled, last_seen]
      properties:
        id:
          type: string
          example: pc-salon
        name:
          type: string
          example: PC Salon
        mac:
          type: string
          pattern: '^([0-9A-Fa-f]{2}[:-]){5}([0-9A-Fa-f]{2})$'
          example: 00:11:22:33:44:55
        status:
          $ref: '#/components/schemas/DeviceStatus'
        ip:
          type: string
          format: ipv4
          example: 192.168.1.50
        ping_enabled:
          type: boolean
          description: Whether the background monitor pings this device
        last_seen:
          type: integer
          format: int64
          description: Unix timestamp of the last up observation, 0 if never
        notes:
          type: string
          maxLength: 500
          description: Free-text note, trimmed on write
        tags:
          type: array
          maxItems: 10
          description: Lowercase slugs ([a-z0-9_-]), normalized and deduped on write
          items:
            type: string
        wake_count:
          type: integer
          minimum: 0
          description: Successful manual and scheduled wakes since registration
        last_wake_at:
          type: integer
          format: int64
          description: Unix timestamp of the last successful wake, absent if never
        monitor_secs:
          type: integer
          minimum: 0
          description: Per-device monitor cadence in seconds, 0 follows the global interval
      additionalProperties: false
    DeviceInput:
      type: object
      required: [id, name, mac]
      description: Same shape as Device. status defaults to unknown when absent
        or empty; PUT ignores body id in favor of the path id.
      properties:
        id:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        mac:
          type: string
          pattern: '^([0-9A-Fa-f]{2}[:-]){5}([0-9A-Fa-f]{2})$'
        status:
          $ref: '#/components/schemas/DeviceStatus'
        ip:
          type: string
          format: ipv4
        ping_enabled:
          type: boolean
        last_seen:
          type: integer
          format: int64
        notes:
          type: string
          maxLength: 500
        tags:
          type: array
          maxItems: 10
          items:
            type: string
        monitor_secs:
          type: integer
          minimum: 0
          description: Per-device monitor cadence in seconds, 0 follows the global interval
      additionalProperties: false
    DeviceStatus:
      type: string
      enum: [up, down, unknown]
    DeviceList:
      type: object
      required: [devices, total, page, per_page]
      properties:
        devices:
          type: array
          items:
            $ref: '#/components/schemas/Device'
        total:
          type: integer
          description: Matches across all pages, before slicing
        page:
          type: integer
        per_page:
          type: integer
      additionalProperties: false
...
    Problem:
      type: object
      description: RFC 9457 problem details. All errors share this envelope.
      required: [type, title, status, detail]
      properties:
        type:
          type: string
          format: uri
          example: about:blank
        title:
          type: string
          example: not found
        status:
          type: integer
          example: 404
        detail:
          type: string
          example: 'storage: device not found'
      additionalProperties: false
    ValidationProblem:
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          required: [errors]
          properties:
            errors:
              type: object
              description: Field to failing-rule map for 422 bodies
              additionalProperties:
                type: string
              example: { "MAC": "mac" }
          additionalProperties: false
  responses:
    NotFound:
      description: Unknown id, RFC 9457 body
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: Duplicate id, RFC 9457 body
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    UnprocessableEntity:
      description: Semantic validation failure, RFC 9457 body
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ValidationProblem'
...
    Unauthorized:
      description: Missing or invalid bearer token, RFC 9457 body
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API token from API_TOKEN. Browsers calling /ws use ?token= instead.